+++ pyproj/transformer.py
@@ -100,7 +100,7 @@ class Transformer(object):
         transformer._transformer = _Transformer.from_pipeline(cstrencode(proj_pipeline))
         return transformer
 
-    def transform(self, xx, yy, zz=None, radians=False):
+    def transform(self, xx, yy, zz=None, radians=False, errcheck=False):
         """
         Transform points between two coordinate systems.
 
@@ -116,6 +116,10 @@ class Transformer(object):
             If True, will expect input data to be in radians and will return radians
             if the projection is geographic. Default is False (degrees). Ignored for
             pipeline transformations.
+        errcheck: boolean, optional (default False)
+            If True an exception is raised if the transformation is invalid.
+            By default errcheck=False and an invalid transformation 
+            returns ``inf`` and no exception is raised.
 
 
         Example:
@@ -148,7 +152,7 @@ class Transformer(object):
         else:
             inz = None
         # call pj_transform.  inx,iny,inz buffers modified in place.
-        self._transformer._transform(inx, iny, inz, radians)
+        self._transformer._transform(inx, iny, inz, radians, errcheck=errcheck)
         # if inputs were lists, tuples or floats, convert back.
         outx = _convertback(xisfloat, xislist, xistuple, inx)
         outy = _convertback(yisfloat, yislist, xistuple, iny)
@@ -158,7 +162,7 @@ class Transformer(object):
         else:
             return outx, outy
 
-    def itransform(self, points, switch=False, radians=False):
+    def itransform(self, points, switch=False, radians=False, errcheck=False):
         """
         Iterator/generator version of the function pyproj.Transformer.transform.
 
@@ -174,6 +178,10 @@ class Transformer(object):
             If True, will expect input data to be in radians and will return radians
             if the projection is geographic. Default is False (degrees). Ignored for
             pipeline transformations.
+        errcheck: boolean, optional (default False)
+            If True an exception is raised if the transformation is invalid.
+            By default errcheck=False and an invalid transformation 
+            returns ``inf`` and no exception is raised.
 
 
         Example:
@@ -218,13 +226,14 @@ class Transformer(object):
             if len(buff) == 0:
                 break
 
-            self._transformer._transform_sequence(stride, buff, switch, radians)
+            self._transformer._transform_sequence(stride, buff, switch,
+                    radians, errcheck=errcheck)
 
             for pt in zip(*([iter(buff)] * stride)):
                 yield pt
 
 
-def transform(p1, p2, x, y, z=None, radians=False):
+def transform(p1, p2, x, y, z=None, radians=False, errcheck=False):
     """
     x2, y2, z2 = transform(p1, p2, x1, y1, z1)
 
@@ -239,6 +248,10 @@ def transform(p1, p2, x, y, z=None, radians=False):
     'radians' is True (default is False), then all input and 
     output coordinates will be in radians instead of the default 
     of degrees for geographic input/output projections.
+    If the optional keyword 'errcheck' is set to True an 
+    exception is raised if the transformation is
+    invalid. By default errcheck=False and ``inf`` is returned for an
+    invalid transformation (and no exception is raised).
 
     In addition to converting between cartographic and geographic
     projection coordinates, this function can take care of datum
@@ -304,10 +317,11 @@ def transform(p1, p2, x, y, z=None, radians=False):
     >>> "%.3f %.3f" % (xr, yr)
     '2.031 0.696'
     """
-    return Transformer.from_proj(p1, p2).transform(x, y, z, radians)
+    return Transformer.from_proj(p1, p2).transform(x, y, z, radians,
+            errcheck=errcheck)
 
 
-def itransform(p1, p2, points, switch=False, radians=False):
+def itransform(p1, p2, points, switch=False, radians=False, errcheck=False):
     """
     points2 = itransform(p1, p2, points1)
     Iterator/generator version of the function pyproj.transform.
@@ -351,4 +365,5 @@ def itransform(p1, p2, points, switch=False, radians=False):
     >>> for pt in itransform(pj, Proj(4326), [(pjx, pjy)], radians=True): '{:.3f} {:.3f}'.format(*pt)
     '2.031 0.696'
     """
-    return Transformer.from_proj(p1, p2).itransform(points, switch, radians)
+    return Transformer.from_proj(p1, p2).itransform(points, switch, radians,
+            errcheck=errcheck)